MAINT: Move matrix_power to linalg
The docstring already assumed it was in linalg, and this ensures
linalg becomes completely independent of matrixlib.
